platform/x86: thinkpad_acpi: Add palm sensor support
BugLink: https://bugs.launchpad.net/bugs/1912033

Add support to thinkpad_acpi for returning the status of the palm
sensor.

This patch builds on the work done previously for the input device
implementation (which was not needed). Both lap and palm sensor are using
sysfs and they are combined into the proxsensor block.

Note: On some platforms, because of an issue in the HW implementation,
the palm sensor presence may be incorrectly advertised as always
enabled even if a palm sensor is not present. The palm sensor is
intended for WWAN transmission power control and should be available
and correct on all WWAN enabled systems. It is not recommended to use
this interface for other use cases.

Signed-off-by: Mark Pearson <markpearson@lenovo.com>
Link: https://lore.kernel.org/r/20201124181154.547518-1-markpearson@lenovo.com
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
(cherry picked from commit 1ac09656bded6e4ea9312631579dd32189e00e62)
Signed-off-by: Alex Hung <alex.hung@canonical.com>
Signed-off-by: Timo Aaltonen <timo.aaltonen@canonical.com>