platform/x86: thinkpad_acpi: correct palmsensor error checking
BugLink: https://bugs.launchpad.net/bugs/1912033

The previous commit adding functionality for the palm sensor had a
mistake which meant the error conditions on initialisation was not checked
correctly. On some older platforms this meant that if the sensor wasn't
available an error would be returned and the driver would fail to load.

This commit corrects the error condition. Many thanks to Mario Oenning
for reporting and determining the issue

Signed-off-by: Mark Pearson <markpearson@lenovo.com>
Link: https://lore.kernel.org/r/20201230024726.7861-1-markpearson@lenovo.com
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
(cherry picked from commit aa44afab87af079e0cf2ead9621d0447798a305e)
Signed-off-by: Alex Hung <alex.hung@canonical.com>
Signed-off-by: Timo Aaltonen <timo.aaltonen@canonical.com>